refactor HFSubset
