+++ mteb/abstasks/AbsTaskClusteringFast.py
@@ -13,8 +13,9 @@ from datasets import Dataset, DatasetDict
 from sklearn.metrics.cluster import v_measure_score
 from torch.utils.data import DataLoader
 
-from mteb.abstasks.TaskMetadata import DescriptiveStatistics, HFSubset
+from mteb.abstasks.TaskMetadata import DescriptiveStatistics
 from mteb.encoder_interface import Encoder
+from mteb.types import HFSubset
 
 from .AbsTask import AbsTask
