+++ mteb/abstasks/AbsTaskRetrieval.py
@@ -9,9 +9,9 @@ from typing import Any, Callable
 
 from datasets import Dataset, DatasetDict
 
-from mteb.abstasks.TaskMetadata import DescriptiveStatistics, HFSubset
+from mteb.abstasks.TaskMetadata import DescriptiveStatistics
 from mteb.encoder_interface import Encoder
-from mteb.types import ScoresDict
+from mteb.types import HFSubset, ScoresDict
 
 from ..evaluation.evaluators import RetrievalEvaluator
 from ..evaluation.evaluators.retrieval_metrics import make_score_dict
