--- mteb/types/metadata.py
@@ -1,3 +1,0 @@
-from __future__ import annotations
-
-HFSubset = str
